refactor: split impact.rs monolith into src/impact/ directory
1522-line monolith → 7 focused modules: types, bfs, hints, analysis,
diff, format, mod. No behavior changes — same public API, all 808 tests
pass. Also adds Code Quality section to roadmap with three identified
improvements.
